feat: optimize UI
