+++ doc/src/provider/LayoutProvider/index.tsx
@@ -1,5 +1,6 @@
 import { SidebarProvider } from 'shadcn-plus';
 import { LeftMenu } from './LeftMenu';
+import { Header } from './Header';
 import { useLocation } from 'react-router-dom';
 
 type IPropsType = {
@@ -8,26 +9,32 @@ type IPropsType = {
 
 export const LayoutProvider = ({ children }: IPropsType) => {
   const { pathname } = useLocation();
-  
-  // Dashboard页面不显示左侧边栏
+
   const isDashboard = pathname === '/';
-  
+
   if (isDashboard) {
-    return <>{children}</>;
-  }
-  
-  // Components页面显示左侧边栏
-  return (
-    <SidebarProvider>
-      <LeftMenu />
-      <div
-        style={{
-          width: 'calc(100% - 256px)',
-          padding: 32,
-        }}
-      >
+    return (
+      <>
+        <Header />
         {children}
-      </div>
-    </SidebarProvider>
+      </>
+    );
+  }
+
+  return (
+    <>
+      {isDashboard ? <Header /> : <></>}
+      <SidebarProvider>
+        <LeftMenu />
+        <div
+          style={{
+            width: 'calc(100% - 256px)',
+            padding: 32,
+          }}
+        >
+          {children}
+        </div>
+      </SidebarProvider>
+    </>
   );
 };
